+++ Sources/MacIMEKit/Args.swift
@@ -148,8 +148,10 @@ public enum ArgsIMED {
    /// Check args
    public static func parse(_ args: [String]) throws -> IMEDState {
       let state = try IMEDState()
-      // Check the first arg
-      if let arg = args.first {
+      // Check the args
+      var i = 0
+      while i < args.count {
+         let arg = args[i]
          switch arg {
          case "--version", "-v":
             IO.out(Defaults.version)
@@ -158,17 +160,19 @@ public enum ArgsIMED {
             IO.out(Help.macimed)
             exit(0)
          case "--log-level", "-l":
-            guard args.count >= 2 else {
+            guard args.count > i + 1 else {
                // throw AppError.imed(.missingLogLevel)
-               IO.err("Missing log level.")
+               IO.err("Missing log level. Use one of debug, info, warn, error.")
                exit(1)
             }
-            let level: LogLevel = Log.getLogLevelByString(args[1])
+            let level: LogLevel = Log.getLogLevelByString(args[i + 1])
             Runtime.logLevel = level
+            i += 1
          default:
             IO.err("Unknown Option: \(arg)")
             exit(1)
          }
+         i += 1
       }
       guard FS.pathExists(state.macimePath ?? "") else {
          throw AppError.imed(.macimeNotFound(state.macimePath ?? ""))